requirements.txt
django
pddiktipy
djangorestframework
markdown       # Markdown support for the browsable API.
django-filter  # Filtering support